+++ src/commands/nns-help.ts
@@ -1,5 +1,4 @@
 import { MessageEmbed } from 'discord.js';
-import { cmdArgsParser } from '../service/utils';
 export class HelpCommand {
     // constructor(){}
     readonly name: string = 'help';
@@ -7,6 +6,8 @@ export class HelpCommand {
     readonly description: string = `
     Display all commands and descriptions
     `;
+    readonly man: string = 'Display all commands and descriptions';
+
     execute(message, args, prefix = message.client.prefix, commands = message.client.commands) {
         const helpEmbed = new MessageEmbed();
         helpEmbed.setTitle('nns-bot Help')