+++ tests/service/commandHandler.spec.ts
@@ -7,4 +7,14 @@ describe('Command Handler', () => {
         const cmdInst = new CommandHandler();
         expect(cmdInst.commandLoader()).to.be.an('array');
     });
+
+    it('command object check.', () => {
+        const cmdInst = new CommandHandler();
+        const cmdList = cmdInst.commandLoader();
+        for (const cmd of cmdList) {
+            for (const key of ['execute', 'name', 'alias', 'description', 'man']) {
+                expect(cmd).to.have.property(key);
+            }
+        }
+    });
 });